requirements.txt
gunicorn
flask
tensorflow
joblib
numpy
Pillow
scikit-learn
